# copilot-demo/src/main/resources/apispec.yaml
...
openapi: 3.0.0
info:
  title: Settlement API
  description: Settlement domain API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.1.0
servers:
  - url: https://localhost:8080/stl/settlement/1.1.0
    description: SwaggerHub API Auto Mocking
...
paths:
  /calendar/{date}:
    get:
      tags:
        - calendars
      summary: the GET method for a single calendar day
      description: Retrieve a calendar day for display or update
      operationId: getSettlementCalendarByDate
      parameters:
        - name: date
          in: path
          description: the unique resource identifier for the document
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: date
            example: 2022-12-31
      responses:
        "200":
          description: single calendar entry
          headers:
            ETag:
              description: an optimistic lock token respresenting the resource state; usually a hex-encoded MD5 hash value
              style: simple
              explode: false
              schema:
                type: string
              example: 43c191bf6d6c3f263a8cd0efd4a058ab
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/calendarDay'
        "403":
          description: Forbidden
        "405":
          description: Method Not Found
    put:
      tags:
        - calendars
      summary: the PUT method for a single calendar day
      description: Updates a calendar day record
      operationId: updateSettlementCalendarDay
      parameters:
        - name: date
          in: path
          description: the unique resource identifier for the document
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: date
            example: 2022-06-03
        - name: If-Match
          in: header
          description: an optimistic lock token respresenting the resource state; usually a hex-encoded MD5 hash value. Obtained from the `Etag` response header of a previous GET request for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 43c191bf6d6c3f263a8cd0efd4a058ab
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/calendarDay'
      responses:
        "200":
          description: calendar entry updated
          headers:
            ETag:
              description: an optimistic lock token respresenting the resource state; usually a hex-encoded MD5 hash value
              style: simple
              explode: false
              schema:
                type: string
              example: 43c191bf6d6c3f263a8cd0efd4a058ab
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/calendarDay'
        "400":
          description: Errors occurred
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/problemFormat'
        "403":
          description: Forbidden
        "409":
          description: Conflict - the `If-Match` value supplied on the request does not match the current `Etag` of the record
  /calendar/next:
    get:
      tags:
        - calendars
      summary: Obtain the next available settlement date
      description: Returns the next settlement date for a given planned date, currency, and payment method.
      operationId: getNextSettlementDate
      parameters:
        - name: paymentMethod
          in: query
          description: The payment method
          required: true
          style: form
          explode: true
          schema:
            type: string
            example: STFO
            enum:
              - STFO
              - SWIFT
        - name: plannedDate
          in: query
          description: The date for the proposed payment
          required: true
          style: form
          explode: true
          schema:
            type: string
            format: date
            example: 2021-12-31
        - name: currency
          in: query
          description: The settlement currency
          required: true
          style: form
          explode: true
          schema:
            $ref: '#/components/schemas/currency'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/nextSettlementDay'
        "400":
          description: Errors occurred
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/problemFormat'
        "403":
          description: Forbidden
...
components:
  schemas:
    calendarDay:
      required:
        - calendarDate
        - currencies
      type: object
      properties:
        calendarDate:
          type: string
          description: The date of this calendar entry
          format: date
          example: 2022-12-31
        currencies:
          maxItems: 14
          minItems: 14
          type: array
          description: There must be one entry for each of the 14 supported currencies
          example: |
            [
              {
                'currency': 'AUD',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'CAD',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'CHF',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'DKK',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'EUR',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'GBP',
                'stfoSettlementAllowed': false,
                'swiftSettlementAllowed': false
              },
              {
                'currency': 'HKD',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'JPY',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'NOK',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'NZD',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'SEK',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'SGD',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'USD',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              },
              {
                'currency': 'ZAR',
                'stfoSettlementAllowed': true,
                'swiftSettlementAllowed': true
              }
            ]
          items:
            $ref: '#/components/schemas/currencyEntry'
        audit:
          $ref: '#/components/schemas/auditBlock'
      description: A day in the settlement calendar. For each supported currency, shows whether or not this day is allowed for settlements.
    currencyEntry:
      required:
        - currency
        - stfoSettlementAllowed
        - swiftSettlementAllowed
      type: object
      properties:
        currency:
          $ref: '#/components/schemas/currency'
        stfoSettlementAllowed:
          type: boolean
          description: Whether or not settlement is allowed in this currency for STFO
          example: true
        swiftSettlementAllowed:
          type: boolean
          description: Whether or not settlement is allowed in this currency for Swift
          example: true
      description: settlement status of the currency
    currency:
      type: string
      description: ISO alpha-3 currency code. There are 14 allowed settlement currencies
      example: NOK
      enum:
        - AUD
        - CAD
        - CHF
        - DKK
        - EUR
        - GBP
        - HKD
        - JPY
        - NOK
        - NZD
        - SEK
        - SGD
        - USD
        - ZAR
    auditBlock:
      required:
        - at
        - user
      type: object
      properties:
        user:
          type: string
          description: The identifier of the user or system user that modified the record
          example: svc99
        at:
          type: string
          description: UTC timestamp of the record modification
          format: date-time
          example: 2021-03-03T08:24:51Z
      description: The data in the audit block is populated directly by the API from the security context when the record is modified
      readOnly: true
    problemFormat:
      type: array
      description: The problem format.
      items:
        $ref: '#/components/schemas/problemFormatItem'
    problemFormatItem:
      required:
        - domain
        - errorCode
        - title
        - type
      type: object
      properties:
        type:
          type: string
          format: uri
          example: https://dxc.com/typerror/premium/2470
        title:
          type: string
          example: title example
        detail:
          type: string
          example: Detailed description of the problem
        domain:
          type: string
          example: PREMIUM
          enum:
            - RISK
            - PREMIUM
            - CLAIM
            - DOCUMENT
            - EXPERTFEES
            - PARTY
            - NOTIFICATION
            - SETTLEMENTS
        errorCode:
          type: integer
          format: int32
          example: 2470
    nextSettlementDay:
      required:
        - currency
        - nextSettlementDate
      type: object
      properties:
        currency:
          $ref: '#/components/schemas/currency'
        nextSettlementDate:
          type: string
          description: The next settlement date for this currency
          format: date
          example: 2022-12-31
      description: Next available settlement date for a currency
...